Relating media to channels
Add media to a channel
List channel media
Delete media from a channel

There's an issue with tests, we can't list a channel's media until it's been
processed. We could fake it, stopping the debugger at the point the video has
been added to the media, and wait for the processing to be over. That would
record a correct cassette in VCR. Thoughts?

Also, our Limelight API credentials have expired.